[Sema] NFC: Move TypeChecker.h to include/swift/Sema
In an effect to make Sema unit-testable, let's start by moving
`TypeChecker.h` to `include/swift/Sema` so it come be referenced
outside of `lib/Sema`.
