+++ lib/Sema/ConstantnessSemaDiagnostics.cpp
@@ -25,11 +25,12 @@
 //===----------------------------------------------------------------------===//
 
 #include "MiscDiagnostics.h"
-#include "TypeChecker.h"
 #include "swift/AST/ASTContext.h"
 #include "swift/AST/ASTWalker.h"
 #include "swift/AST/ParameterList.h"
 #include "swift/AST/SemanticAttrs.h"
+#include "swift/Sema/TypeChecker.h"
+
 using namespace swift;
 
 /// Check whether a given \p decl has a @_semantics attribute with the given
