+++ lib/Sema/DebuggerTestingTransform.cpp
@@ -26,9 +26,8 @@
 #include "swift/AST/ParameterList.h"
 #include "swift/AST/SourceFile.h"
 #include "swift/AST/Stmt.h"
+#include "swift/Sema/TypeChecker.h"
 #include "swift/Subsystems.h"
-
-#include "TypeChecker.h"
 
 using namespace swift;
