+++ lib/Sema/TypeCheckAttr.cpp
@@ -17,7 +17,6 @@
 #include "MiscDiagnostics.h"
 #include "TypeCheckObjC.h"
 #include "TypeCheckType.h"
-#include "TypeChecker.h"
 #include "swift/AST/ASTVisitor.h"
 #include "swift/AST/ClangModuleLoader.h"
 #include "swift/AST/DiagnosticsParse.h"
@@ -35,6 +34,7 @@
 #include "swift/AST/Types.h"
 #include "swift/Parse/Lexer.h"
 #include "swift/Sema/IDETypeChecking.h"
+#include "swift/Sema/TypeChecker.h"
 #include "clang/Basic/CharInfo.h"
 #include "llvm/ADT/STLExtras.h"
 #include "llvm/Support/Debug.h"
