+++ lib/Sema/TypeCheckCodeCompletion.cpp
@@ -16,7 +16,6 @@
 
 #include "swift/Subsystems.h"
 #include "ConstraintSystem.h"
-#include "TypeChecker.h"
 #include "TypeCheckObjC.h"
 #include "TypeCheckType.h"
 #include "CodeSynthesis.h"
@@ -42,6 +41,7 @@
 #include "swift/Parse/Lexer.h"
 #include "swift/Sema/IDETypeChecking.h"
 #include "swift/Sema/CodeCompletionTypeChecking.h"
+#include "swift/Sema/TypeChecker.h"
 #include "swift/Strings.h"
 #include "llvm/ADT/DenseMap.h"
 #include "llvm/ADT/PointerUnion.h"
