+++ lib/Sema/TypeCheckConstraints.cpp
@@ -19,7 +19,6 @@
 #include "ConstraintSystem.h"
 #include "MiscDiagnostics.h"
 #include "SolutionResult.h"
-#include "TypeChecker.h"
 #include "swift/AST/ASTVisitor.h"
 #include "swift/AST/ASTWalker.h"
 #include "swift/AST/DiagnosticSuppression.h"
@@ -30,6 +29,7 @@
 #include "swift/AST/TypeCheckRequests.h"
 #include "swift/Basic/Statistic.h"
 #include "swift/Sema/CodeCompletionTypeChecking.h"
+#include "swift/Sema/TypeChecker.h"
 #include "llvm/ADT/DenseMap.h"
 #include "llvm/ADT/SmallPtrSet.h"
 #include "llvm/ADT/SmallVector.h"
