+++ lib/Sema/TypeCheckDecl.cpp
@@ -17,7 +17,6 @@
 #include "CodeSynthesis.h"
 #include "ConstraintSystem.h"
 #include "DerivedConformances.h"
-#include "TypeChecker.h"
 #include "TypeCheckAccess.h"
 #include "TypeCheckAvailability.h"
 #include "TypeCheckConcurrency.h"
@@ -46,6 +45,7 @@
 #include "swift/Parse/Lexer.h"
 #include "swift/Parse/Parser.h"
 #include "swift/Sema/IDETypeChecking.h"
+#include "swift/Sema/TypeChecker.h"
 #include "swift/Serialization/SerializedModuleLoader.h"
 #include "swift/Strings.h"
 #include "swift/AST/NameLookupRequests.h"
