+++ lib/Sema/TypeCheckDeclPrimary.cpp
@@ -19,7 +19,6 @@
 #include "CodeSynthesis.h"
 #include "ConstraintSystem.h"
 #include "DerivedConformances.h"
-#include "TypeChecker.h"
 #include "TypeCheckAccess.h"
 #include "TypeCheckDecl.h"
 #include "TypeCheckAvailability.h"
@@ -46,6 +45,7 @@
 #include "swift/Basic/Statistic.h"
 #include "swift/Parse/Lexer.h"
 #include "swift/Parse/Parser.h"
+#include "swift/Sema/TypeChecker.h"
 #include "swift/Serialization/SerializedModuleLoader.h"
 #include "swift/Strings.h"
 #include "swift/AST/NameLookupRequests.h"
