+++ lib/Sema/TypeCheckSwitchStmt.cpp
@@ -14,14 +14,14 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "TypeChecker.h"
 #include "swift/AST/ASTPrinter.h"
 #include "swift/AST/DiagnosticsSema.h"
 #include "swift/AST/Pattern.h"
 #include "swift/Basic/Debug.h"
 #include "swift/Basic/STLExtras.h"
-
 #include "swift/Basic/APIntMap.h"
+#include "swift/Sema/TypeChecker.h"
+
 #include <llvm/ADT/APFloat.h>
 
 #include <forward_list>
