+++ lib/Sema/TypeChecker.cpp
@@ -17,7 +17,6 @@
 
 #include "swift/Subsystems.h"
 #include "ConstraintSystem.h"
-#include "TypeChecker.h"
 #include "TypeCheckObjC.h"
 #include "TypeCheckType.h"
 #include "CodeSynthesis.h"
@@ -40,6 +39,7 @@
 #include "swift/Basic/STLExtras.h"
 #include "swift/Parse/Lexer.h"
 #include "swift/Sema/IDETypeChecking.h"
+#include "swift/Sema/TypeChecker.h"
 #include "swift/Strings.h"
 #include "llvm/ADT/DenseMap.h"
 #include "llvm/ADT/PointerUnion.h"
